fix(bindings): rename main bindings package
