+++ .gitignore
@@ -20,3 +20,4 @@ go.work.sum
 
 build
 .idea
+.vscode/
